MAP_POPULATE for hoard
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@5591 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ lib/hoard-38/src/heaplayers/zoneheap.h
@@ -58,7 +58,7 @@ namespace HL {
 
   private:
 
-    inline static size_t align (int sz) {
+    inline static size_t align (size_t sz) {
       return (sz + (sizeof(double) - 1)) & ~(sizeof(double) - 1);
     }
